new retry_task_map is ready to be used
