+++ tests/test_otaclient_common/test_retry_task_map.py
@@ -18,12 +18,10 @@ from __future__ import annotations
 import logging
 import random
 import time
-from functools import partial
 
 import pytest
 
-from otaclient_common.common import get_backoff
-from otaclient_common.retry_task_map import RetryTaskMap, RetryTaskMapInterrupted
+from otaclient_common import retry_task_map
 
 logger = logging.getLogger(__name__)
 
@@ -36,7 +34,6 @@ class TestRetryTaskMap:
     WAIT_CONST = 100_000_000
     TASKS_COUNT = 2000
     MAX_CONCURRENT = 128
-    DOWNLOAD_GROUP_NO_SUCCESS_RETRY_TIMEOUT = 6  # seconds
     MAX_WAIT_BEFORE_SUCCESS = 10
 
     @pytest.fixture(autouse=True)
@@ -64,81 +61,28 @@ class TestRetryTaskMap:
         self._succeeded_tasks[idx] = True
         return idx
 
-    def test_retry_keep_failing_timeout(self):
-        _keep_failing_timer = time.time()
-        with pytest.raises(RetryTaskMapInterrupted):
-            _mapper = RetryTaskMap(
-                backoff_func=partial(get_backoff, factor=0.1, _max=1),
-                max_concurrent=self.MAX_CONCURRENT,
-                max_retry=0,  # we are testing keep failing timeout here
-            )
-            for done_task in _mapper.map(
+    def test_retry_exceed_retry_limit(self):
+        with retry_task_map.ThreadPoolExecutorWithRetry(
+            max_total_retry=200
+        ) as executor:
+            for _ in executor.ensure_tasks(
                 self.workload_aways_failed, range(self.TASKS_COUNT)
             ):
-                if not done_task.fut.exception():
-                    # reset the failing timer on one succeeded task
-                    _keep_failing_timer = time.time()
-                    continue
-                if (
-                    time.time() - _keep_failing_timer
-                    > self.DOWNLOAD_GROUP_NO_SUCCESS_RETRY_TIMEOUT
-                ):
-                    logger.error(
-                        f"RetryTaskMap successfully failed after keep failing in {self.DOWNLOAD_GROUP_NO_SUCCESS_RETRY_TIMEOUT}s"
-                    )
-                    _mapper.shutdown(raise_last_exc=True)
-
-    def test_retry_exceed_retry_limit(self):
-        with pytest.raises(RetryTaskMapInterrupted):
-            _mapper = RetryTaskMap(
-                backoff_func=partial(get_backoff, factor=0.1, _max=1),
-                max_concurrent=self.MAX_CONCURRENT,
-                max_retry=3,
-            )
-            for _ in _mapper.map(self.workload_aways_failed, range(self.TASKS_COUNT)):
                 pass
+        assert executor.executor_interrupted
 
     def test_retry_finally_succeeded(self):
-        _keep_failing_timer = time.time()
-
-        _mapper = RetryTaskMap(
-            backoff_func=partial(get_backoff, factor=0.1, _max=1),
-            max_concurrent=self.MAX_CONCURRENT,
-            max_retry=0,  # we are testing keep failing timeout here
-        )
-        for done_task in _mapper.map(
-            self.workload_failed_and_then_succeed, range(self.TASKS_COUNT)
-        ):
-            # task successfully finished
-            if not done_task.fut.exception():
-                # reset the failing timer on one succeeded task
-                _keep_failing_timer = time.time()
-                continue
-
-            if (
-                time.time() - _keep_failing_timer
-                > self.DOWNLOAD_GROUP_NO_SUCCESS_RETRY_TIMEOUT
+        with retry_task_map.ThreadPoolExecutorWithRetry() as executor:
+            for _ in executor.ensure_tasks(
+                self.workload_failed_and_then_succeed, range(self.TASKS_COUNT)
             ):
-                _mapper.shutdown(raise_last_exc=True)
+                pass
         assert all(self._succeeded_tasks)
 
     def test_succeeded_in_one_try(self):
-        _keep_failing_timer = time.time()
-        _mapper = RetryTaskMap(
-            backoff_func=partial(get_backoff, factor=0.1, _max=1),
-            max_concurrent=self.MAX_CONCURRENT,
-            max_retry=0,  # we are testing keep failing timeout here
-        )
-        for done_task in _mapper.map(self.workload_succeed, range(self.TASKS_COUNT)):
-            # task successfully finished
-            if not done_task.fut.exception():
-                # reset the failing timer on one succeeded task
-                _keep_failing_timer = time.time()
-                continue
-
-            if (
-                time.time() - _keep_failing_timer
-                > self.DOWNLOAD_GROUP_NO_SUCCESS_RETRY_TIMEOUT
+        with retry_task_map.ThreadPoolExecutorWithRetry() as executor:
+            for _ in executor.ensure_tasks(
+                self.workload_succeed, range(self.TASKS_COUNT)
             ):
-                _mapper.shutdown(raise_last_exc=True)
+                pass
         assert all(self._succeeded_tasks)
